refactor: fix wrong onboarding folders organization
