+++ CarGurus App/lib/data/data_sources/onboarding/onboarding_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:carguru/domain/models/onboarding_model.dart';
+import 'package:carguru/domain/models/onboarding/onboarding_model.dart';
 import 'package:injectable/injectable.dart';
 import 'onboarding_data_source_impl.dart';
 import '../../../core/utils/results.dart';
