+++ CarGurus App/lib/data/data_sources/onboarding/onboarding_data_source_impl.dart
@@ -1,6 +1,6 @@
 import 'package:carguru/core/utils/results.dart';
 import 'package:carguru/data/data_sources/onboarding/onboarding_data_source.dart';
-import 'package:carguru/domain/models/onboarding_model.dart';
+import 'package:carguru/domain/models/onboarding/onboarding_model.dart';
 
 import '../../../utils/App_content.dart';
