+++ CarGurus App/lib/domain/models/onboarding/onboarding_model.dart
