+++ CarGurus App/lib/domain/repositories/onboarding/onboarding_repository.dart
@@ -1,6 +1,6 @@
 import 'package:carguru/data/data_sources/onboarding/onboarding_data_source.dart';
 import 'package:carguru/data/repositories/onboarding/onboarding_repository_impl.dart';
-import 'package:carguru/domain/models/onboarding_model.dart';
+import 'package:carguru/domain/models/onboarding/onboarding_model.dart';
 import 'package:injectable/injectable.dart';
 
 import '../../../core/utils/results.dart';
