+++ CarGurus App/lib/domain/use_cases/onboarding/get_onboarding_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:carguru/domain/models/onboarding_model.dart';
+import 'package:carguru/domain/models/onboarding/onboarding_model.dart';
 import 'package:carguru/domain/repositories/onboarding/onboarding_repository.dart';
 import 'package:injectable/injectable.dart';
